+++ Assets/UniGLTF/Runtime/UniGLTF/IO/AltTask/AsyncMethodBuilderAttribute.cs
@@ -1,0 +1,12 @@
+namespace System.Runtime.CompilerServices
+{
+    public sealed class AsyncMethodBuilderAttribute : Attribute
+    {
+        public AsyncMethodBuilderAttribute(Type builderType)
+        {
+            BuilderType = builderType;
+        }
+
+        public Type BuilderType { get; }
+    }
+}
