+++ Assets/UniGLTF/Runtime/UniGLTF/IO/AltTask/TaskQueue.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6a65e00e71613c24e865e35173c30be0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
